+++ src/main/java/com/kpei/mkbd/datatransform/dto/Vd58Dto.java
@@ -2,11 +2,13 @@ package com.kpei.mkbd.datatransform.dto;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Setter
 @Getter
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
 public class Vd58Dto extends Vd5Base {
-    private double nilai;
+    private BigDecimal nilai;
 }
